requirements.txt
nba_api>=1.4.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pandas>=2.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
asyncpg>=0.29.0
openai>=1.12.0
pydantic-settings>=2.1.0
tiktoken>=0.6.0
feedparser>=6.0.0
trafilatura>=1.8.0
apscheduler>=3.10.0
